pyproject.toml
[project]
name = "vllm-srv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=0.20.0",
    "chromadb==1.1.1",
    "faiss-cpu==1.12.0",
    "fastapi>=0.100.0",
    "langchain==0.3.27",
    "langchain-chroma==0.2.6",
    "langchain-community==0.3.30",
    "langchain-core==0.3.78",
    "langchain-openai==0.3.35",
    "langchain-qdrant==0.2.1",
    "langchain-text-splitters==0.3.11",
    "langgraph>=1.0.1",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-vector-stores-chroma>=0.5.5",
    "llama-index-vector-stores-faiss>=0.5.2",
    "llama-index-vector-stores-qdrant>=0.8.8",
    "llama-index[all]>=0.14.12",
    "numpy>=1.21.0",
    "packaging>=21.0",
    "protobuf>=3.20.0",
    "psutil>=5.9.0",
    "pydantic>=2.0.0",
    "pymupdf==1.26.7",
    "qdrant-client<1.16",
    "ray>=2.5.0",
    "safetensors>=0.3.0",
    "sentencepiece>=0.1.99",
    "tokenizers>=0.13.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0,!=4.57.0",
    "typesense==1.1.1",
    "typing-extensions>=4.0.0",
    "uvicorn[standard]>=0.20.0",
    "vllm>=0.6.0",
]
